Broke agents into their own file. Fixed the recentMessages selection loop inverting order of messages.
